Adding expense to Room done
1) Added Repository for income expense
2) Added shared view model for sharing data between activity and fragment

+++ app/src/main/java/com/n2ksp/expense_tracker/data/room/IncomeExpenseDBModel.kt
@@ -22,7 +22,7 @@ class IncomeExpenseDBModel {
     var memo: String? = ""
 
     @ColumnInfo(name = "amount")
-    var amount: Int = 0
+    var amount: Float = 0.0f
 
     @ColumnInfo(name = "date")
     var date = Date()
